Create a ComputationalBurdenAnalyzer Class.
