+++ src/simulations/optimisation.py
@@ -847,21 +847,6 @@ class Optimiser():
 
         return res
 
-    def _save_computational_burden_opti(
-        self, time_to_solve_opti, number_opti_constraints):
-        """Save computational burden results to file."""
-        if os.path.exists(f"{self.paths['record_folder']}/computational_res.npz"):
-            computational_res = np.load(f"{self.paths['record_folder']}/computational_res.npz")
-            opti_timer = np.append(
-                computational_res['opti_timer'], time_to_solve_opti)
-            n_constraints = np.append(
-                computational_res['n_constraints'], number_opti_constraints)
-            np.savez_compressed(f"{self.paths['record_folder']}/computational_res.npz",
-                opti_timer=opti_timer, n_constraints=n_constraints)
-        else:
-            np.savez_compressed(f"{self.paths['record_folder']}/computational_res.npz",
-                opti_timer=time_to_solve_opti, n_constraints=number_opti_constraints)
-
     def plot_results(self, res, prm, folder=None):
         """Plot the optimisation results for homes."""
         store = res['store']
